+++ jocasta/collector.py
@@ -54,11 +54,11 @@ def setup_connectors(config):
     for name, section in config.items():
         args = convert_config_stanza(section)
         if name == 'file_system':
-            connectors['file_system'] = file_system.FileSystemConnector(**args)
-        elif name == 'adafruit':
-            connectors['adafruit'] = io_adafruit.IOAdafruitConnector(**args)
+            connectors[name] = file_system.FileSystemConnector(**args)
+        elif name == 'io_adafruit':
+            connectors[name] = io_adafruit.IOAdafruitConnector(**args)
         elif name == 'influxdb':
-            connectors['influxdb'] = influx.InfluxDBConnector(**args)
+            connectors[name] = influx.InfluxDBConnector(**args)
     return connectors