+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.2.0]
 ### Fixed
 - Error when missing temperature range in config file.
+
+### Added
+- InfluxDB connector
+- Tests
+
+### Changed
+- `adafruit` to `io_adafruit` in the config.
 
 ## [0.1.0] - 2020-08-11
 ### Added